+++ packages/polaris-viz-core/src/utilities/index.ts
@@ -1,5 +1,4 @@
-export {createTheme} from './createThemes';
-export {createThemes} from './createThemes';
+export {createTheme, createThemes} from './createThemes';
 export {createGradient} from './createGradient';
 export {getSeriesColors} from './getSeriesColors';
 export {getFilteredSeries} from './getFilteredSeries';
@@ -7,3 +6,4 @@ export {uniqueId} from './uniqueId';
 export {isGradientType} from './isGradientType';
 export {curveStepRounded} from './curveStepRounded';
 export {getAnimationTrail} from './getAnimationTrail';
+export {paddingStringToObject} from './paddingStringToObject';